[AllocPartition, Clang] Implement TypeHashPointerSplit mode
Implement the TypeHashPointerSplit mode: This mode assigns a partition
ID based on the hash of the allocated type's name, where the top half
ID-space is reserved for types that contain pointers and the bottom half
for types that do not contain pointers.

This mode with max partitions of 2 (`-fsanitize-alloc-partition-max=2`)
may also be valuable for heap hardening strategies that simply separate
pointer types from non-pointer types.

Make it the new default mode.

Link: https://discourse.llvm.org/t/rfc-a-framework-for-allocator-partitioning-hints/87434